Record parses through the umbrella conftest so the suites still collect together
